Split constants into a common module
