+++ common/src/main/java/de/guntram/mcmod/antighost/Constants.java
@@ -1,0 +1,12 @@
+package de.guntram.mcmod.antighost;
+
+public class Constants {
+    public static final String MOD_ID = "antighost";
+    public static final String MOD_NAME = "AntiGhost";
+    // The x, y, and z range to refresh blocks at
+    public static final int REFRESH_RANGE = 4;
+    // The delay between executes
+    public static final int DELAY = 10;
+    public static final String REFRESH_DESCRIPTION_KEY = "key." + MOD_ID + ".reveal";
+    public static final String CONTROLS_CATEGORY_KEY = "key.categories." + MOD_ID;
+}
