+++ common/src/main/resources/Antighost.xcf
