+++ common/src/main/resources/assets/antighost/lang/cs_cz.json
