+++ common/src/main/resources/assets/antighost/lang/es_es.json
