+++ common/src/main/resources/assets/antighost/lang/es_mx.json
