+++ common/src/main/resources/assets/antighost/lang/fr_fr.json
