+++ common/src/main/resources/assets/antighost/lang/ja_jp.json
