+++ common/src/main/resources/assets/antighost/lang/pl_pl.json
