+++ common/src/main/resources/assets/antighost/lang/pt_pt.json
