+++ common/src/main/resources/assets/antighost/lang/zh_cn.json
