+++ common/src/main/resources/assets/antighost/lang/zh_tw.json
