+++ forge/src/main/resources/META-INF/mods.toml
@@ -23,7 +23,7 @@ displayName="AntiGhost" #mandatory
 # A URL for the "homepage" for this mod, displayed in the mod UI
 displayURL="http://github.com/gbl/AntiGhost/" #optional
 # A file name (in the root of the mod JAR) containing a logo for display
-logoFile="Antighost.png" #optional
+logoFile= "../Antighost.png" #optional
 # A text field displayed in the mod UI
 # credits="" #optional
 # A text field displayed in the mod UI
