+++ settings.gradle
@@ -1,0 +1,4 @@
+rootProject.name = "antighost"
+
+include "common"
+include "forge"
